feat: criando paginas custom de login e registro
