+++ src/app/create-account/[[...sign-up]]/page.tsx
@@ -1,0 +1,9 @@
+import { SignUp } from "@clerk/nextjs";
+
+export default function CreateAccountPage(){
+    return(
+        <div className="h-screen flex justify-center items-center flex-col">
+            <SignUp />
+        </div>
+    )
+}
