+++ src/app/login/[[...sign-in]]/page.tsx
@@ -1,0 +1,9 @@
+import { SignIn } from "@clerk/nextjs";
+
+export default function LoginPage(){
+    return(
+        <div className="h-screen flex justify-center items-center flex-col">
+            <SignIn />
+        </div>
+    )
+}
